+++ EasyAssertions/Assertions/ObjectAssertions.cs
@@ -8,14 +8,17 @@ namespace EasyAssertions;
 /// </summary>
 public static class ObjectAssertions
 {
+    /// <summary>Disallowed comparison.</summary>
+    [Obsolete("Use ShouldBeValue instead", true)]
+    public static void ShouldBe<TActual>(this TActual? actual, TActual expected, string? message = null) where TActual : struct { }
+
     /// <summary>
     /// Asserts that two objects are equal, using the default equality comparer.
     /// </summary>
-    public static Actual<TActual> ShouldBe<TActual, TExpected>([NotNull] this TActual? actual, TExpected expected, string? message = null)
+    public static Actual<TActual> ShouldBe<TActual, TExpected>(this TActual actual, TExpected expected, string? message = null)
         where TExpected : TActual
-        where TActual : class
     {
-        return actual.RegisterNotNullAssertion(c =>
+        return actual.RegisterAssertion(c =>
             {
                 if (!c.Test.ObjectsAreEqual(actual, expected))
                 {
@@ -30,7 +33,7 @@ public static class ObjectAssertions
     /// <summary>
     /// Asserts that two values are equal, using the default equality comparer.
     /// </summary>
-    public static Actual<TActual> ShouldBe<TActual>(this TActual actual, TActual expected, string? message = null)
+    public static Actual<TActual> ShouldBeValue<TActual>(this TActual actual, TActual expected, string? message = null)
         where TActual : struct
     {
         actual.RegisterAssertion(c =>
@@ -45,19 +48,14 @@ public static class ObjectAssertions
     /// <summary>
     /// Asserts that a nullable value is equal to another value, using the default equality comparer.
     /// </summary>
-    public static Actual<TActual> ShouldBe<TActual>([NotNull] this TActual? actual, TActual expected, string? message = null)
+    public static Actual<TActual?> ShouldBeValue<TActual>(this TActual? actual, TActual? expected, string? message = null)
         where TActual : struct
     {
-        actual.RegisterAssertion(c =>
+        return actual.RegisterAssertion(c =>
             {
-                if (!actual.HasValue)
-                    throw c.StandardError.NotEqual(expected, actual, message);
-
-                if (!c.Test.ObjectsAreEqual(actual.Value, expected))
+                if (!c.Test.ObjectsAreEqual(actual, expected))
                     throw c.StandardError.NotEqual(expected, actual, message);
             });
-
-        return new Actual<TActual>(actual!.Value);
     }
 
     /// <summary>
@@ -65,7 +63,6 @@ public static class ObjectAssertions
     /// </summary>
     public static Actual<TActual> ShouldNotBe<TActual, TNotExpected>(this TActual actual, TNotExpected notExpected, string? message = null)
         where TNotExpected : TActual
-        where TActual : class?
     {
         return actual.RegisterAssertion(c =>
             {
@@ -77,7 +74,20 @@ public static class ObjectAssertions
     /// <summary>
     /// Asserts that two vales are not equal, using the default equality comparer.
     /// </summary>
-    public static Actual<TActual> ShouldNotBe<TActual>(this TActual actual, TActual notExpected, string? message = null)
+    public static Actual<TActual> ShouldNotBeValue<TActual>(this TActual actual, TActual notExpected, string? message = null)
+        where TActual : struct
+    {
+        return actual.RegisterAssertion(c =>
+            {
+                if (c.Test.ObjectsAreEqual(actual, notExpected))
+                    throw c.StandardError.AreEqual(notExpected, actual, message);
+            });
+    }
+
+    /// <summary>
+    /// Asserts that two vales are not equal, using the default equality comparer.
+    /// </summary>
+    public static Actual<TActual?> ShouldNotBeValue<TActual>(this TActual? actual, TActual? notExpected, string? message = null)
         where TActual : struct
     {
         return actual.RegisterAssertion(c =>
@@ -117,13 +127,27 @@ public static class ObjectAssertions
     /// Asserts that the given object is not a null reference.
     /// </summary>
     public static Actual<TActual> ShouldNotBeNull<TActual>([NotNull] this TActual? actual, string? message = null)
-        where TActual : notnull
+        where TActual : class
     {
         return actual.RegisterNotNullAssertion(c =>
             {
                 if (Equals(actual, null))
                     throw c.StandardError.IsNull(message);
             });
+    }
+
+    /// <summary>
+    /// Asserts that the given object is not a null reference.
+    /// </summary>
+    public static Actual<TActual> ShouldNotBeNull<TActual>([NotNull] this TActual? actual, string? message = null)
+        where TActual : struct
+    {
+        actual.RegisterAssertion(c =>
+            {
+                if (Equals(actual, null))
+                    throw c.StandardError.IsNull(message);
+            });
+        return new Actual<TActual>(actual!.Value);
     }
 
     /// <summary>